style : interactions 관심사 분리
import 상대경로로 수정
