+++ repo/interactions/urls.py
@@ -1,10 +1,5 @@
-from django.urls import path
-
-from repo.interactions.views import *
+from django.urls import include, path
 
 urlpatterns = [
-    path("follow/", FollowListAPIView.as_view(), name="my_follow_list"),
-    path("<int:id>/follow/", FollowListCreateDeleteAPIView.as_view(), name="follow"),
-    path("block/", BlockListAPIView.as_view(), name="my_block_list"),
-    path("<int:id>/block/", BlockListCreateDeleteAPIView.as_view(), name="block"),
+    path("relationship/", include("repo.interactions.relationship.urls")),
 ]
