+++ repo/profiles/management/commands/seed_user_and_details.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 from django_seed import Seed
 from faker import Faker
 
-from repo.interactions.models import Relationship
+from repo.interactions.relationship.models import Relationship
 from repo.profiles.models import CustomUser, UserDetail
